[UPDATE] Version 1.1.0 (#97)
* [FIX] Fastlane metadata (#84)

* fix fastlane metadata

* fix fastlane metadata

* [FIX] F-droid (#86)

* fix fastlane metadata

* fix fastlane metadata

* remove useless properties

* add ndkVersion

* [CI] Try to homogenize F-Droid and Github Actions (#89)

* update github actions to java 21

* enable minify

* changelog

* create beta release

* beta ressources

* lines

* [UPDATE] Version code

* [UPDATE] Version name

* [UPDATE] Add dependenciesInfo in gradle

* [FEATURE]  Access to unknown files (#94)

* Create Other artwork

* Revert "Create Other artwork"

This reverts commit 029650711a102fffe0be58adf75c75b36bbc5526.

* Create Unknown artwork

* flag isUnknown

* get unknown files

* backup entry point for unknown

* UnknownCategory

* backup UnknownScreen

* backup files

* get duration for unknown + limitedDispatcher

* backup UnknownViewModel

* rename mediaId to artworkId

* backup UnknownScreen

* backup UnknownScreen

* strings

* backup UnknownScreenContent

* remove FixedChip in ArtworkDescription

* MediaThumbnail

* navigate to UnknownScreen

* backup UnknownItem

* backup UnknownItem

* backup UnknownItem

* UnknownItem

* UnknownItem

* getUnknownMediaFrom

* unknownMedia in PlayerViewModel

* MediaMockups.episodesWithStatus

* SearchViewModel

* HomeViewModelTest

* backup

* backup UnknownViewModelTest

* improve FakeArtworkRepository

* UnknownViewModelTest

* update version

* [FEATURE] Optional TMDB Token (#96)

* dont request token

* text button

* token_desc_3

* backup Token SnackBar

* HomeSnackbar

* HomeSnackbar

* TokenScreen changes

* starting screen fix

* speed up

* just show a message after token validation

* try to retrieve unknown Medias

* fetch correctly unknown medias

* sort unknown medias

* imports

* remove clearToken

* getUnknownMedias if no token

* clear token

* clear token

* fix

* fix tests

* version and changelog

* strings

* token not null but blank

* string

+++ app/src/main/java/com/mskd/flux/data/repository/catalog/CatalogRepositoryImpl.kt
@@ -7,6 +7,8 @@ import com.mskd.flux.data.source.media.MediaSource
 import com.mskd.flux.data.source.media.MediaSourceTMDBImpl.Companion.TAG
 import com.mskd.flux.model.UserFile
 import com.mskd.flux.model.artwork.Artwork
+import com.mskd.flux.model.artwork.Episode
+import com.mskd.flux.model.artwork.Movie
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -21,13 +23,11 @@ class CatalogRepositoryImpl @Inject constructor(
     private val mediaSourceLocal: MediaSource,
     private val mediaSourceTmdb: MediaSource,
     private val db: DatabaseDao,
-    scope: CoroutineScope
+    private val scope: CoroutineScope
 ) : CatalogRepository {
 
     private val _catalogFlow = MutableStateFlow(CatalogRepository.State())
     override val flow: StateFlow<CatalogRepository.State> = _catalogFlow.asStateFlow()
-
-    private var isSyncing = false
 
     init {
         scope.launch {
@@ -44,59 +44,60 @@ class CatalogRepositoryImpl @Inject constructor(
 
             // Update content
             _catalogFlow.update { content ->
-                content.copy(
-                    isLoading = isSyncing,
-                    artworks = medias.sortedBy { it.title }
-                )
+                content.copy(artworks = medias.sortedBy { it.title })
             }
 
         }
     }
 
-    override suspend fun syncCatalog() {
+    override fun syncCatalog() {
 
-        isSyncing = true
+        if (_catalogFlow.value.isLoading)
+            return
 
-        _catalogFlow.update { it.copy(isLoading = true) }
+        scope.launch {
 
-        try {
+            _catalogFlow.update { it.copy(isLoading = true) }
 
-            // Fetch all files, local and online (if possible)
-            val allFiles = getFiles()
-            val dbFileNames = db.getAllFileNames()
+            tryToRetrieveUnknownMedias()
 
-            // Delete medias with missing files
-            db.deleteMediasWithNoFiles(allFiles)
+            try {
 
-            // Get new medias from TMBD
-            val newFiles = allFiles.filter { !dbFileNames.contains(it.name) }
-            val (newArtworks, newMovies, newEpisodes) = mediaSourceTmdb.getMedias(files = newFiles)
+                // Fetch all files, local and online (if possible)
+                val allFiles = getFiles()
+                val dbFileNames = db.getAllFileNames()
 
-            // Save new medias
-            db.insertArtworks(newArtworks)
-            db.insertMovies(newMovies)
-            db.insertEpisodes(newEpisodes)
+                // Delete medias with missing files
+                db.deleteMediasWithNoFiles(allFiles)
 
-            val allArtworks = db.getArtworks()
+                // Get new medias from TMBD
+                val newFiles = allFiles.filter { !dbFileNames.contains(it.name) }
+                val (newArtworks, newMovies, newEpisodes) = mediaSourceTmdb.getMedias(files = newFiles)
 
-            // Update content
-            _catalogFlow.update { content ->
-                content.copy(
-                    isLoading = false,
-                    artworks = allArtworks.sortedBy { it.title }
-                )
+                // Save new medias
+                db.insertArtworks(newArtworks)
+                db.insertMovies(newMovies)
+                db.insertEpisodes(newEpisodes)
+
+                val allArtworks = db.getArtworks()
+
+                // Update content
+                _catalogFlow.update { content ->
+                    content.copy(
+                        isLoading = false,
+                        artworks = allArtworks.sortedBy { it.title }
+                    )
+                }
+
+            } catch (e: Exception) {
+                Log.e(TAG, "[syncCatalog] Fail to sync catalog", e)
+
+                _catalogFlow.update { content ->
+                    content.copy(isLoading = false)
+                }
             }
 
-        } catch (e: Exception) {
-            Log.e(TAG, "[syncCatalog] Fail to sync catalog", e)
-
-            _catalogFlow.update { content ->
-                content.copy(isLoading = false)
-            }
         }
-
-        isSyncing = false
-
     }
 
     override suspend fun getFiles() : List<UserFile> {
@@ -114,6 +115,62 @@ class CatalogRepositoryImpl @Inject constructor(
         }
 
         return localFiles
+
+    }
+
+    private suspend fun tryToRetrieveUnknownMedias() {
+
+        try {
+
+            val unknownMedias = db.getUnknownMedias()
+            val files = unknownMedias.map { it.file }
+
+            val (newArtworks, newMovies, newEpisodes) = mediaSourceTmdb.getMedias(files = files)
+
+            val moviesToSave = arrayListOf<Movie>()
+            val episodesToSave = arrayListOf<Episode>()
+            val mediasToDelete = arrayListOf<Episode>()
+
+            newMovies.forEach { movie ->
+
+                unknownMedias.find { it.file == movie.file }?.let { unknownMedia ->
+
+                    val newMovie = movie.copy(
+                        currentTime = unknownMedia.currentTime,
+                        status = unknownMedia.status
+                    )
+
+                    moviesToSave.add(newMovie)
+                    mediasToDelete.add(unknownMedia)
+
+                }
+
+            }
+
+            newEpisodes.filter { !it.isUnknown }.forEach { episode ->
+
+                unknownMedias.find { it.file == episode.file }?.let { unknownMedia ->
+
+                    val newEpisode = episode.copy(
+                        currentTime = unknownMedia.currentTime,
+                        status = unknownMedia.status
+                    )
+
+                    episodesToSave.add(newEpisode)
+                    mediasToDelete.add(unknownMedia)
+
+                }
+
+            }
+
+            db.insertArtworks(newArtworks.filter { !it.isUnknown })
+            db.insertMovies(moviesToSave)
+            db.insertEpisodes(episodesToSave)
+            db.deleteEpisodes(mediasToDelete)
+
+        } catch (e: Exception) {
+            Log.e(TAG, "Fail to retrieve unknown medias", e)
+        }
 
     }
